Preserve the datatype
Currently, when arguments are generated in trace files all of them result in `String` data type, so every time I try to match source and destinations (rank is int, but arguments are in string), I had to change the type. This will preserve the type from original MPI calls. 

If you think this will be extra/inefficient here, please ignore it. I will attend to them in the match code.