timeline: Correct timestamps
Current timestamps are 10× larger than expected.

Timestamps from Recorder are in seconds and are converted with the timer
resolution of the file. Which defaults to 1e-7. Chrome traces expect
microseconds as timestamp, optional with a nanosecond fraction.

Ignore timer resolution from Recorder and just convert to microseconds
and format them as double with precision 3.